remove out of place section from update.graphql
This section shows the reader how to link the Lambda function to a GraphQL schema in a "technically correct" way, then goes on to describe a much more accurate AppSync schema to connect to the custom resolver. I was tripped up by this a few days ago until I read a little further and found the "real" schema.